Moved to nestjs Added dev docker-compose Added prod docker-compose
