+++ .env.example
@@ -1,0 +1,6 @@
+VALIDATOR=0xa7cea918946b15a8b02f8af93992cfc012d8f88677b6c2785aeac1d7b2f13366e47a8aab2a8850be0375d7851742a1ab
+CRON_PATTERN=0 0 * * * *
+INFLUX_HOST=influx
+INFLUX_PORT=8086
+INFLUX_DB=etherstaker
+INFLUX_MEASUREMENT=performance
