+++ docker-compose-dev.yml
@@ -1,0 +1,12 @@
+version: "3"
+
+services:
+  validator-revenue:
+    build: .
+    command: npm run start:dev
+    volumes:
+      - .:/src
+    env_file: .env
+
+  influx:
+    image: influxdb:alpine
